requirements.txt
lxml==4.6.4
Pillow==8.4.0
python-docx==0.8.11
python-pptx==0.6.21
XlsxWriter==3.0.2
